+++ stream_cqsa/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 from .stable_stream import (
     stream_cqsa_forward,